handling empty extractor ids + schema refactors
