+++ src/types/expected-outcome.ts
@@ -1,5 +1,7 @@
 export type {
   EvaluationSource,
+  EvaluationSourceExtractor,
+  EvaluationSourceExtractors,
   ExpectedOutcomeMode,
   ExpectedOutcomeSchemaField,
   ExpectedOutcomeSchema,
